After running the link checker for a couple of weeks we identified a problem with AustLII (the Australian legal repository)  throwing 404s for every request. Further investigation  identified that their automated tools had flagged the tool as a web scraper, hence the errors.
AustLII pointed out that the link checker uses a GET call but we don't need to, because all we care about is the resonse code, not the actual page content. They requested a change to use a HEAD request, and add a unique and identifiable User Agent name instead of using the generic requests module name. This both avoids being miscategorised as a web scraping tool (since a HEAD request doesn't touch the body of the pages), and presents users of the tool as good web citizens by clearly identifying what is sending the request.

I've made these changes locally with success, but it seems that this is likely to be an issue with other publishers and the suggested changes were pretty sensible so I have put together a small commit and suggest it be pulled into the main repository.